Changed task names and removed all hardcoded references to them
Task names are now always the same for each playable faction.
This makes things much easier to manage and it was an unnecessary complication as to two factions can be active at the same time.
All script references to task names are through the PatrolManager and therefore defined in one place.
A few references exist outside the scripts, namely on infil/exfil MissionTriggers and in the multitasks themselves.

If multiple factions are ever required to be active at same time then the infil/exfil triggers will simply be required to be faction specific (still up in the air).
The multitasks themselves are already faction specific so names would just need updating.